requirements.txt
pyautogui
Pillow
numpy
opencv-python